Updated weapons

+++ gam666/fwk4gps/Game/GameObjects/WeaponSpread.h
@@ -7,7 +7,7 @@ class Player;
 class WeaponSpread : public Weapon {
 public:
 	void fireProjectile();
-	WeaponSpread(Player* o, int cdDuration, int mHeat, int hPerShot);
+	WeaponSpread(Player* o, float cdDuration, int mHeat, int hPerShot);
 	~WeaponSpread();
 };
